# Build performance enhancement