--- a/build/webpack.browser.js
+++ b/build/webpack.browser.js
@@ -12,7 +12,7 @@ const webpack = require('webpack');
 const ExtractTextPlugin = require("extract-text-webpack-plugin");
 
 module.exports = ({
-  devtool = 'source-map',
+  devtool = 'cheap-module-eval-source-map',
   entry = '.js',
   outputPath = path.resolve(__dirname, 'dist'),
   publicPath = './dist/',