--- a/build/webpack.browser.js
+++ b/build/webpack.browser.js
@@ -39,7 +39,7 @@ module.exports = ({
     loaders: [
       {
         test: /\.jsx|js?$/,
-        loader: 'babel-loader',
+        loader: 'babel-loader?cacheDirectory=true',
         exclude: /node_modules/
       },
       {