--- a/build/webpack.static.js
+++ b/build/webpack.static.js
@@ -43,7 +43,7 @@ module.exports = ({
     loaders: [
       {
         test: /\.(jsx|js)$/,
-        loader: 'babel-loader',
+        loader: 'babel-loader?cacheDirectory=true',
         exclude: /node_modules/
       },
       {