See #95 for source issue

## Description
* Switches the webpack.config.js to exporting a factory function, thus making dev builds not inherit production settings
* Turns on babel caching (saves 3-5s build time)
* Changed devtool values based on reading an article (saves ~1s build time)
* Disabled uglifying tests in prod builds (saves ~10 sec)
* Disabled uglify sourcemaps (saves ~10 sec)

## Motivation and Context
Build times were ~38 seconds.
Pre-uglify build times were ~11 sec.
Now a dev build is ~7.5sec, and prod build ~16 sec.

## How Has This Been Tested?
`npm run build`
`npm run production`

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)