@krambuhl 2016-11-08
super duper!

seeing ~6s build times down from ~8.5s on my machine